requirements.txt
python-decouple==3.8
colorama==0.4.6